Refactor physical ops continued
